+++ src/main/java/com/starkbank/error/StarkBankError.java
@@ -1,0 +1,8 @@
+package com.starkbank.error;
+
+public abstract class StarkBankError extends RuntimeException {
+
+    public StarkBankError(String message) {
+        super(message);
+    }
+}