+++ src/main/java/com/starkbank/error/UnknownError.java
@@ -1,8 +1,8 @@
 package com.starkbank.error;
 
-public final class UnknownError extends RuntimeException{
+public final class UnknownError extends StarkBankError {
 
-    public UnknownError(String message){
+    public UnknownError(String message) {
         super(message);
     }
 }